Add files via upload
This uses the IAC Driver to send gametrak data as controller data. 
E.g. This will allow you to use the gametrak as MIDI controller in Ableton
If you want to use this with Windows, you will need to install something like LoopMIDI - https://www.tobias-erichsen.de/software/loopmidi.html

Note that you will need to have the gametrak.maxpat downloaded and in the same folder as this patch to use it. You also need to plug in the gametrak before opening the patch.